changes to deal with mediawiki's CORS breaking the presidents demo
Exhibit doesn't set CORS headers, which means that it can't directly
access images cross site for the map view (because it needs to tinker
with those images in canvases that become "tainted" by the cross site
request).  So, the painter service is invoked for cross origin image
requests.  Some recent changes to mediawiki commons mean that exhibit
painter is failing to fetch those images properly, probably because
*painter* is not using the right CORS headers.  So, we're getting
markers without images.  A quick fix is to modify Exhibit to use cors
headers, which enable it to fetch images directly without the
painter.  However, CORS is complicated and introducing these CORS
headers may have other negative effects.  We need to investigate.